Version 0.7 BETA (1G0) * Added the ability to teleport to broken decoys! * Added command for searching the broken-decoy log! * Added some testing functionality. * Updated Metrics.

+++ DecoyBlocks/src/com/fishrock123/decoyblocks/DBBlockListener.java
@@ -22,7 +22,7 @@ public class DBBlockListener implements Listener {
 	
 	@EventHandler (priority = EventPriority.LOWEST)
 	public void onBlockBreak(BlockBreakEvent e) {
-		if (database.decoyLocations.containsKey(e.getBlock().getLocation()) && e.getPlayer().hasPermission("decoyblocks.decoy")) {
+		if (database.decoyLocations.containsKey(e.getBlock().getLocation()) && e.getPlayer().hasPermission("decoyblocks.decoy") && !database.testList.contains(e.getPlayer())) {
 			database.remove(e.getBlock());
 			e.getPlayer().sendMessage("Removed " + new DBBlock(e.getBlock()).toString() + " from the decoyList.");
 			return;
